release: 修复 v2.2.0 Release 构建并关闭 Web 资源打包
- 修复 Kotlin 编译错误（Compose Scope、协程上下文、委托属性智能转换等）
- 移除 buildFeatures.compose，避免与 Kotlin 2.x Compose Compiler 插件冲突
- 修正 baseline-prof.txt 与 proguard-rules.pro 的 R8 兼容格式
- 确认 versionName=2.2.0 / versionCode=20200，Web 资源通过 sourceSets 排除

+++ app/src/main/java/com/silas/omaster/ui/features/CameraXViewfinderScreen.kt
@@ -3,6 +3,7 @@ package com.silas.omaster.ui.features
 import android.Manifest
 import android.content.pm.PackageManager
 import android.graphics.Bitmap
+import android.graphics.SurfaceTexture
 import android.net.Uri
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -42,6 +43,8 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Cancel
+import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material.icons.filled.Cameraswitch
 import androidx.compose.material.icons.filled.FlashAuto
 import androidx.compose.material.icons.filled.FlashOff
@@ -50,6 +53,8 @@ import androidx.compose.material.icons.filled.PhotoCamera
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -73,6 +78,7 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.graphicsLayer
@@ -194,7 +200,7 @@ fun CameraXViewfinderScreen(
                 zoomRatio = currentZoom,
                 iso = proParams.iso,
                 shutterSpeedNs = proParams.shutterSpeedNs,
-                exposureCompensation = proParams.exposureCompensation ?: 0f
+                exposureCompensation = (proParams.exposureCompensation ?: 0f).toFloat()
             ),
             gyroscopeStable = gyroscopeStable,
             upperBrightnessRatio = brightnessStats.upperBrightnessRatio,
